hand2print/mobile_backend/requirements.txt
albumentations==1.3.0
numpy==1.23.5
opencv_python_headless==4.7.0.72
pyclipper==1.3.0.post4
pyctcdecode==0.5.0
Shapely==2.0.1
torch==2.0.0
torchvision==0.15.1
fastapi==0.100.0
pydantic==2.0.2
uvicorn==0.22.0
python-multipart==0.0.6
Pillow==10.0.0
firebase-admin==6.2.0
